requirements.txt
Flask
mysql-connector-python
google-generativeai
anthropic
numpy
pandas
pgmpy